test(chat): split matrixChat.spec.ts by theme
At 21 tests and 7 minutes on chromium, this was by far the heaviest spec file
in the suite. CI packs whole spec files into shards, so a single file is a floor
on how short a shard can be, and this one alone was longer than the target shard
duration.

Split along the seams that were already there: rooms and messages, encryption,
folders, session verification. The hooks are duplicated into each file, which is
what the sibling matrixChat*.spec.ts files already do.
